fix: skip build/push image if it already exists (#108)
When building the Helm chart, we do not need to build/push images that
already exist. By skipping that, we can save up to ~1h of build time (if
we sum up all the durations in all the build/push jobs) and ~10min of
waiting (the maximum duration is ~10min, for the github_hooks
application) on the "Generate Helm chart" promotion.