fix: handle jobs with no name (#137)
Currently, we are not validating that required fields for jobs are
always present. That can lead to a scenario where a job is created
without a name – `sem create` – which will break listing jobs.

This pull request fixes that by:
- Validating that required fields are present when creating/updating job
records
- Defaulting to empty string when serializing jobs without a name, to
ensure we are handling the old empty-name jobs still in the database and
being returned when listing jobs using the API.