fix(github_hooks): raise unknown error on unhandled octokit exception (#152)
When we receive an error from Octokit that we are not currently
handling, the `pull_request` method on the repo host client returns nil,
which can lead to a `undefined method `[]' for nil` error on the hook
handler. See [this
comment](https://github.com/renderedtext/tasks/issues/7601#issuecomment-2729445192)
for more details. Instead, to help us troubleshoot issues, we should log the actual error
message that we are receiving from GitHub.