fix(repohub): GetFiles should fetch Git updates from origin only for specified branch (#159)
Right now, when fetching the YAML files for workflow editor, we fetch
Git remote updates from all branches in the repository. That can take a
while to run for big repositories, which can lead to 500s in workflow
editor page. Here, we limit the `git fetch origin` command to fetch
changes only for the branch we are interested.