feat(artifacthub): unwrap directories when listing (#186)
Use the new `unwrap_directories` field (added in
https://github.com/renderedtext/internal_api/pull/835) in the
artifacthub internal API to control the `WrapSubDirectories` behavior in
the `storage.ListOptions`. Currently, we hardcoded it to always wrap
directories when listing, but in some cases, we need all the files,
regardless of them being directly under the path specified, or under one
of its subdirectories.

### Other changes

I had to update google.golang.org/grpc to v1.71.0 due to the protobuf
re-generation.