feat(front): use a job to fetch yamls in workflow editor (#206)
The workflow editor currently fetches yaml files in a synchronous call
before the page is rendered, which is very unstable for large
repositories and can lead to 500 errors.
This PR changes this behavior so the workflow editor starts with a zero
state screen with the spinner, and the job is started in the background
to fetch yaml files asynchronously.
Once the files are fetched successfully, the workflow editor is fully
rendered with visual and text editor components.
This behavior will remain behind a feature flag until we verify that it
works as expected at scale.

Co-authored-by: Lucas Pinheiro <lpinheiro@renderedtext.com>