fix(ci): pin versions of Go tools and scan only Dockerfile misconfigurations (#208)
This pull request fixes a few issues that are making our daily builds
fail:
- The newest version of revive requires Go 1.23, which fails the build
for some applications which are still on Go 1.22 -
[job](https://semaphore.semaphoreci.com/jobs/4a1e5113-c36c-45ed-9514-059510015678#L557).
It's always a good practice to pin the versions of the tools used in CI,
so we do that here.
- For some applications, we are not building the image before checking
it with trivy, which causes the CI job to fail -
[job](https://semaphore.semaphoreci.com/jobs/a23a9e73-1a76-4143-b1a4-197e778ed9b6)
- We are currently ignoring the severity in the `trivy config` command,
which causes some LOW misconfigurations to fail builds. Also, those LOW
misconfigurations come from Kubernetes YAMLs, which is a newer thing
that trivy can do. Since our goal with using trivy config was to scan
for misconfigurations in Dockerfiles, we should not scan Kubernetes
templates at all for now, so we update the security toolbox to be more
direct with what we want to scan in `trivy config`

Ref: https://github.com/renderedtext/tasks/issues/7804