Update TOS.md (#375)
Adding a distinction between the Community Edition, which is licensed
under the Apache 2.0 License, and the Enterprise Edition, which is
governed by our proprietary Enterprise License Agreement. The Terms of
Service now clarify which license applies depending on the edition in
use.

Additionally, the “ Restrictions” section has been updated to reflect
these differences.

## 📝 Description
<!-- Describe your changes in detail -->

## ✅ Checklist
- [ ] I have tested this change
- [ ] This change requires documentation update