feat(v1alpha api): Allow passing parameters in Run Workflow API call (#323)
Changes the Run workflow API implementation to allow passing parameters
and improve performance.

Previously, the request was synchronously waiting for repo-proxy to
create a hook by retrieving data from the git provider and then
schedule the workflow on the plumber.
This had performance issues, and it did not support passing parameters to
be used in the pipeline.

The new implementation goes directly to the plumber, which only saves the
request in the database before responding, and all hook processing is
done asynchronously.
This should improve performance, and also the plumber supports passing
parameters out-of-the-box.