fix(api v1alpha): Use pipeline_file as param name in Run wf API (#420)
In the recent changes to how the Run workflow API works in the backend,
we changed by mistake the expected name for the parameter that holds the
name of the pipeline file. This PR fixes that issue and adds a test to
verify that the request is properly formed based on the input
parameters.