fix: Several typos (#422) (#474)
## 📝 Description
Wanted to have a look around the codebase, wow, there's a lot going on,
gotta say, it's impressive.

Spotted a couple of typoes in my travels so ran a bit of spellchecking
and amended a few of them, tbh they're all readable enough, just turbo
nitpicking, let me know/remove bits if this'll cause issues.

N.b. Did this before in the Semaphore docs,
https://github.com/semaphoreci/semaphore/pull/290 and it's nice to use
docs + tools where you don't just have to moan if something doesn't make
sense/spot a typo, but being able to fix it is a delight.

## ✅ Checklist
- [ ] I have tested this change
- [ ] This change requires documentation update

---------

Co-authored-by: Pieter Oliver <68863060+pieterocp@users.noreply.github.com>
Co-authored-by: Veljko Maksimovic <45179708+VeljkoMaksimovic@users.noreply.github.com>