Update Documentation typos.md (#662)
Documentation update > typos

## 📝 Description

Correcting typos:

https://docs.semaphore.io/using-semaphore/organizations#init-agent 

Some types of pipelines require initialization job to evaluate and
compile the them before the workload beings.
Some types of pipelines require initialization job to evaluate and
compile them before the workload begins.

Semaphore tries to pick the best type of agent automatically but you
change it for the whole organization. This is particularly useful when
you are using self-hosted agents.
Semaphore tries to pick the best type of agent automatically, but you
can change it for the whole organization. This is particularly useful
when you are using self-hosted agents.

## ✅ Checklist
- [ ] I have tested this change
- [x] This change requires documentation update