feat(github_hooks): track repository changes (#752)
## 📝 Description

This is the first step to update the repository settings after the
change is done in GitHub.
We will be able to change the repository name and the default branch.
For now, we are not listening to repository events, so we are utilizing
the push webhooks. Because of that, the change will be visible only
after the next push after the repository is renamed/changed.

We are focusing on GitHub OAuth Token-based projects; we will fix the
GitHub App later.

https://github.com/renderedtext/tasks/issues/8836

## ✅ Checklist
- [x] I have tested this change
- [x] ~This change requires documentation update~