feat(repository): implement remote repository change consumer (#772)
## 📝 Description

This PR implements `RemoteRepositoryChangedConsumer` in RepositoryHub,
which is now refreshing repository data, and it's switched from remote
repository ID to Semaphore repository ID.

In GitHubHooks, we are now emitting `RemoteRepositoryChanged` every time
we detect that a webhook commit is from a repository that has changed
its default branch or name from what we have saved in the database.

This PR fixes the problem with changing the repository name on GitHub.
It supports only projects created with a GitHub OAuth token. Other
providers, including GitHub App, will be added in further PRs.

## ✅ Checklist
- [x] I have tested this change
- [x] ~This change requires documentation update~